# openapi.yaml
openapi: 3.1.0
info:
  title: Ticketing System API
  version: 1.0.0
  description: |
    REST API for the Ticketing System — a microservices-based event ticketing platform.

    All requests are made through the **API Gateway** at `http://localhost:8000`.
    The gateway handles JWT verification and forwards requests to the appropriate
    downstream service.

    ## Authentication
    Protected endpoints require a JWT bearer token obtained from `POST /auth/login`.
    Include it in the `Authorization` header:
    ```
    Authorization: Bearer <access_token>
    ```

    ## Services
    | Tag | Upstream Service | Description |
    |-----|-----------------|-------------|
    | auth | auth-service | User registration, login, token verification |
    | booking | booking-service | Events, venues, seat reservations, payments |
    | status | booking-status-service | Reservation and booking queries |
    | payment | payment-mock | Mock payment provider (testing only) |
servers:
  - url: http://localhost:8000
    description: Local development (API Gateway)
security: []
...
paths:
...
  /status/bookings/{booking_id}:
    get:
      tags: [status]
      summary: Get confirmed booking details
      description: |
        Retrieve the details of a confirmed booking. A booking is created
        automatically when a payment capture succeeds (via webhook).

        Only the authenticated user who owns the booking may view it.

        **Requires authentication.**
      operationId: getBookingDetails
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/bookingIdPath"
      responses:
        "200":
          description: Booking details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookingDetails"
        "401":
          description: Missing or invalid JWT token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Authenticated user does not own this booking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                detail: Access denied
        "404":
          description: Booking not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                detail: Booking not found
  /status/user/bookings:
    get:
      tags: [status]
      summary: Get current user's bookings
      description: |
        Return a paginated list of all bookings belonging to the authenticated user,
        sorted by creation date descending (most recent first).

        **Requires authentication.**
      operationId: getUserBookings
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            default: 50
            minimum: 1
            maximum: 200
          required: false
          description: Maximum number of bookings to return.
        - in: query
          name: offset
          schema:
            type: integer
            default: 0
            minimum: 0
          required: false
          description: Pagination offset (number of records to skip).
      responses:
        "200":
          description: Paginated list of user bookings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserBookingsResponse"
        "401":
          description: Missing or invalid JWT token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ─────────────────────────────────────────────────────────────
  # Payment Mock (testing only)
  # ─────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from `POST /auth/login`.
  parameters:
...
    bookingIdPath:
      in: path
      name: booking_id
      required: true
      schema:
        type: string
        format: uuid
      description: UUID of the confirmed booking.
  schemas:
...
    ErrorResponse:
      type: object
      required: [detail]
      properties:
        detail:
          type: string
          description: Human-readable error description.
          example: "Resource not found"

    # ── Auth ────────────────────────────────────────────────────
...
    BookingDetails:
      type: object
      required: [booking_id, event_id, event_name, user_email, status, seats, payment_status, created_at, updated_at]
      properties:
        booking_id:
          type: string
          format: uuid
          description: Unique confirmed-booking identifier.
        event_id:
          type: string
          example: "evt-abc123"
        event_name:
          type: string
          description: Human-readable event name resolved from the events table.
          example: "Rock Concert 2026"
        event_start_time:
          type: string
          format: date-time
          nullable: true
          description: Start time of the linked event.
          example: "2026-06-15T19:00:00"
        user_email:
          type: string
          format: email
          example: user@example.com
        status:
          type: string
          enum: [reserved, confirmed, expired, cancelled]
          example: confirmed
        seats:
          type: array
          items:
            type: string
          example: ["A1", "A2"]
        payment_status:
          type: string
          description: Status from the payment provider.
          example: completed
        total_amount:
          type: number
          nullable: true
          description: Total amount charged. `null` if not yet captured.
          example: 150.00
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        expires_at:
          type: string
          format: date-time
          nullable: true
          description: Reservation expiry time. Populated while status is `reserved`.
    UserBookingsResponse:
      type: object
      required: [bookings, total]
      properties:
        bookings:
          type: array
          items:
            $ref: "#/components/schemas/BookingDetails"
        total:
          type: integer
          description: Total number of bookings for the user (before pagination).
          example: 3

    # ── Payment Mock ─────────────────────────────────────────────
...